improve

+++ bds.sh
@@ -1,45 +1,49 @@
 #!/data/data/com.termux/files/usr/bin/bash
 
 install_bds() {
-    # 安装必要软件包
-    pkg install proot wget jq -y
+    if [ ! -d "$HOME/BDS/ubuntu-rootfs" ]; then
+        # 安装必要软件包
+        pkg install proot wget jq -y
 
-    # 创建 BDS 文件夹
-    cd ~
-    mkdir BDS && cd BDS
+        # 创建 BDS 文件夹
+        cd ~
+        mkdir BDS && cd BDS
 
-    # 下载 Ubuntu 镜像
-    echo "正在下载 Ubuntu 镜像..."
-    git clone https://jihulab.com/hmsjy2017/ubuntu-rootfs
-    cd ubuntu-rootfs
+        # 下载 Ubuntu 镜像
+        echo "正在下载 Ubuntu 镜像..."
+        git clone https://jihulab.com/hmsjy2017/ubuntu-rootfs
+        cd ubuntu-rootfs
 
-    # 解压 Ubuntu 镜像
-    echo "正在解压 Ubuntu 镜像..."
-    proot --link2symlink tar -xvJf ubuntu-rootfs.tar.xz --exclude="dev"||:
-    rm ~/BDS/ubuntu-rootfs/ubuntu-rootfs.tar.gz
+        # 解压 Ubuntu 镜像
+        echo "正在解压 Ubuntu 镜像..."
+        proot --link2symlink tar -xJf ubuntu-rootfs.tar.xz --exclude="dev"||:
+        rm ~/BDS/ubuntu-rootfs/ubuntu-rootfs.tar.xz
 
-    # 下载启动脚本并赋予执行权限
-    cd ..
-    wget https://raw.gitmirror.com/hmsjy2017/bdsmgr/main/start-ubuntu.sh
-    chmod +x start-ubuntu.sh
-    wget https://raw.gitmirror.com/hmsjy2017/bdsmgr/main/startbds
-    chmod +x startbds
-    echo "Ubuntu 容器已安装完成"
+        # 下载启动脚本并赋予执行权限
+        cd ..
+        wget https://raw.gitmirror.com/hmsjy2017/bdsmgr/main/start-ubuntu.sh
+        chmod +x start-ubuntu.sh
+        wget https://raw.gitmirror.com/hmsjy2017/bdsmgr/main/startbds -O $HOME/BDS/ubuntu-rootfs/usr/bin/startbds
+        chmod +x ./ubuntu-rootfs/usr/bin/startbds
+        echo "Ubuntu 容器已安装完成"
 
-    # 在容器中下载并安装Minecraft基岩版
-    echo "正在安装 BDS..."
-    cd ~/BDS/ubuntu-rootfs/root/
-    mkdir mc && cd mc
-    LATEST_VERSION=`curl -s https://raw.gitmirror.com/Bedrock-OSS/BDS-Versions/main/versions.json | jq -r '.linux.stable'`
-    echo "Linux BDS 最新版： [${LATEST_VERSION}]"
-    DOWNLOAD_URL=`curl -s https://raw.gitmirror.com/Bedrock-OSS/BDS-Versions/main/linux/${LATEST_VERSION}.json | jq -r '.download_url'`
-    echo "正在下载最新版 BDS 压缩包..."
-    wget ${DOWNLOAD_URL}
-    unzip bedrock-server-${LATEST_VERSION}.zip
-    rm bedrock-server-${LATEST_VERSION}.zip
-    chmod +x bedrock_server
+        # 在容器中下载并安装 Minecraft 基岩版
+        echo "正在安装 BDS..."
+        cd ~/BDS/ubuntu-rootfs/root/
+        mkdir mc && cd mc
+        LATEST_VERSION=`curl -s https://raw.gitmirror.com/Bedrock-OSS/BDS-Versions/main/versions.json | jq -r '.linux.stable'`
+        echo "Linux BDS 最新版： [${LATEST_VERSION}]"
+        DOWNLOAD_URL=`curl -s https://raw.gitmirror.com/Bedrock-OSS/BDS-Versions/main/linux/${LATEST_VERSION}.json | jq -r '.download_url'`
+        echo "正在下载最新版 BDS 压缩包..."
+        wget ${DOWNLOAD_URL}
+        unzip bedrock-server-${LATEST_VERSION}.zip
+        rm bedrock-server-${LATEST_VERSION}.zip
+        chmod +x bedrock_server
 
-    echo "BDS 已在 Ubuntu 容器中安装完成。"
+        echo "BDS 已在 Ubuntu 容器中安装完成。"
+    else
+        echo "检测到已安装 BDS，退出"
+    fi
     exit
 }
 
@@ -47,6 +51,8 @@ run_bds() {
     if [ ! -d "$HOME/BDS/ubuntu-rootfs" ]; then
         echo "请先安装 BDS。"
     else
+        echo "正在进入容器..."
+        echo "请在容器中执行 startbds 命令启动 BDS，需要停止 BDS 时执行 stop 命令 。"
         cd ~/BDS
         ./start-ubuntu.sh
     fi
